requirements.txt
# tkinter is part of Python standard library
# Add any additional dependencies here
pyinstaller>=5.0.0
Pillow>=9.0.0
